refactor(gateway): reorganize gateway imports and add new server implementation
